refactor : rename folder name
