+++ src/components/quote-carousel/Quote.tsx
