+++ src/components/quote-carousel/QuoteCarousel/QuoteCarousel.tsx
