+++ src/components/quote-carousel/QuoteCarousel/QuoteSlider.test.tsx
